refactor: move normalize_sensor_id from UDP adapter to domain/sensor
normalize_sensor_id is a sensor identity concern (MAC hex normalization),
not a UDP transport concern. Moving it to domain/sensor.py eliminates
layer violations where infra/config and HTTP adapters reached into the
UDP protocol adapter for a generic utility.

The function is now self-contained in domain with no protocol dependency.
The definition is removed from protocol.py entirely (no-backward-compat
policy).

Closes #740
