+++ apps/server/vibesensor/adapters/http/_helpers.py
@@ -7,7 +7,7 @@ from contextlib import contextmanager
 
 from fastapi import HTTPException
 
-from vibesensor.adapters.udp.protocol import normalize_sensor_id
+from vibesensor.domain import normalize_sensor_id
 from vibesensor.shared.exceptions import (
     AnalysisNotReadyError,
     DataCorruptError,
